The torrent.py script now works very well in all instances, plus stderr and stdout were swapped.
I swapped stderr and stdout on the basis that subprocess.Popen will catch exceptions on stderr, and it's probably a good idea to let the user know about errors and whatnot. So now script events will occur on stdout.

The torrent.py script now supports turning turtle mode on and off, changing the turtle mode download limit, as well as listing current downloads and their progress.