Well, where to begin. New name, deluge script added and more.
I've added a half-done script for supporting deluge. It's a bit dodgy
and only some stuff works but it's a good start. Config now supports
comments and will automatically remove them.

Not sure if it was mentioned in another commit, but a group of Thread
classes has been added to ensure that if the SMTP client crashes, all
current emails will be saved, and the server will then reattempt to
connect to the SMTP server and subsquently send the emails. Currently it
runs in the ProcessThreadsStarter thread, after checking for emails, it
will try and reconnect. Problems WILL occur though if more than two
ProcessThreadsStarters are running at once so this should be rectified.

Aside from that, as of this commit the software is now called makeme.