Fixed Config().save bug, added custom messages and help for scripts.
The Config().save() method had a bug where files weren't closed, nothing
major but worth fixing.

Custom messages have been added, so the introductory email can be
changed for people deploying the software that want to manage it
themselves and such. Introductory messages and the help messages are
currently customisable.

A new help script was added into the scripts for getting help with using
all the different scripts. Subsequently all the scripts have been given
a help option now, so in the event scripts are run with the body
argument being help, they will instead display help messages and quit.