Default arguments to settings, server specification too.
I've added the ability to use default settings if a particular setting
isn't found in the config file, which is very nice. This means much
tidier config files!

Also in this commit is the ability for users to specify other servers as
well, so if they want to use hotmail, or yahoo or whatever they can.
It's completely untested and will probably remain that way until someone
tries it out as I'm perfectly content using gmail for this.