Mostly a documentation and bug fix commit. Fixed some other small things.
The patterns for scripting were moved out of the for loop and compiled.
There was a bug in the ProcessThreadsStarter.run method that meant it
would check over and over again, crash or not.

I also added some locking into the method as well, and removed
threading.iterate() to replace it with a more localised version.