Locking during message retrieval is better now, fixed bug for responses.
During message retrieval, a lock was acquired for the entire block. This
meant that if it had trouble retrieving emails, it could go for
potentially 2 and a half minutes locked up. This has been fixed so that
they will only happen for 30 seconds a time now. Still not good, but
much better.

In the last commit with all my testing, I forgot to uncomment the code
that replies to emails. This has now been rectified.