Lots of bug fixes!
Just little bug fixes here and there regarding lots of things. Files
weren't currently processed for attaching to outgoing emails.

Unsent emails will now be saved to a user specified file, specified
using the unsent_save_location tag in config files.

There was a bug in the MPD script, for random songs that are currently
playing, the title may not be retrievable for whatever reason. This is
caught and ignored when it occurs.

Emails that contained script processing didn't specify who sent the
emails they're sending back, this is now fixed.

Whenever locking occurs now in code that may bomb out with exceptions,
the lock is ALWAYS released in a finally block. This way we know it will
always be run.