Security, proper coding practice and bug fixes.
This commit is pretty big! I've added in security for storing
the user's password. It's not the most secure thing in the
world but it's much better than storing it plaintext. So for a
user to change the account that's used, they must use the
set-user.py script. If anyone has serious gripes with my
security, they can implement something better as I openly
admit I'm not a security expert.

I've completely removed my config class and associated functions to ma
I've completely removed my config class and associated functions because
there's no point reinventing the wheel. The configparser module included
in the standard libraries is more then capable, so that's what I'll be
using from now on! As a result the use of default config options is much
more clean now.